Allow inline blocks to extend concrete classes
Given we don't want the original concrete class to be redefined, we subclass
it before evaluating the inline block on its subclass.

However, getting this to work was trickier than I anticipated, because
the class instance variables we were using on the section DSL configuration
would not be inherited by subclasses.

After experimenting with a variety of approaches, I decided to
pull active support's class_attribute extension. It's a shame that
this brings in a bunch of other dependencies, but for now this seemed
like the best solution.

I also had to deal with the confusion around test_ids when using
a concrete class. I decided the concrete class test id will never
be used, unless when the class is being explicitly .loaded.
(ref: https://github.com/oysterhr/tabasco/issues/17).

This feels more consistent, and foster code reuse.